usm: Move storeProtocolStats to the OS specific files
