+++ pkg/network/state.go
@@ -680,70 +680,6 @@ func (ns *networkState) storeDNSStats(stats dns.StatsByKeyByNameByType) {
 	}
 }
 
-// storeHTTPStats stores the latest HTTP stats for all clients
-func (ns *networkState) storeHTTPStats(allStats map[http.Key]*http.RequestStats) {
-	storeUSMStats[http.Key, *http.RequestStats](
-		allStats,
-		ns.clients,
-		func(c *client) map[http.Key]*http.RequestStats { return c.httpStatsDelta },
-		func(c *client, m map[http.Key]*http.RequestStats) { c.httpStatsDelta = m },
-		func(prev, new *http.RequestStats) { prev.CombineWith(new) },
-		ns.maxHTTPStats,
-		stateTelemetry.httpStatsDropped.Inc,
-	)
-}
-
-func (ns *networkState) storeHTTP2Stats(allStats map[http.Key]*http.RequestStats) {
-	storeUSMStats[http.Key, *http.RequestStats](
-		allStats,
-		ns.clients,
-		func(c *client) map[http.Key]*http.RequestStats { return c.http2StatsDelta },
-		func(c *client, m map[http.Key]*http.RequestStats) { c.http2StatsDelta = m },
-		func(prev, new *http.RequestStats) { prev.CombineWith(new) },
-		ns.maxHTTPStats,
-		stateTelemetry.http2StatsDropped.Inc,
-	)
-}
-
-// storeKafkaStats stores the latest Kafka stats for all clients
-func (ns *networkState) storeKafkaStats(allStats map[kafka.Key]*kafka.RequestStats) {
-	storeUSMStats[kafka.Key, *kafka.RequestStats](
-		allStats,
-		ns.clients,
-		func(c *client) map[kafka.Key]*kafka.RequestStats { return c.kafkaStatsDelta },
-		func(c *client, m map[kafka.Key]*kafka.RequestStats) { c.kafkaStatsDelta = m },
-		func(prev, new *kafka.RequestStats) { prev.CombineWith(new) },
-		ns.maxKafkaStats,
-		stateTelemetry.kafkaStatsDropped.Inc,
-	)
-}
-
-// storePostgresStats stores the latest Postgres stats for all clients
-func (ns *networkState) storePostgresStats(allStats map[postgres.Key]*postgres.RequestStat) {
-	storeUSMStats[postgres.Key, *postgres.RequestStat](
-		allStats,
-		ns.clients,
-		func(c *client) map[postgres.Key]*postgres.RequestStat { return c.postgresStatsDelta },
-		func(c *client, m map[postgres.Key]*postgres.RequestStat) { c.postgresStatsDelta = m },
-		func(prev, new *postgres.RequestStat) { prev.CombineWith(new) },
-		ns.maxPostgresStats,
-		stateTelemetry.postgresStatsDropped.Inc,
-	)
-}
-
-// storeRedisStats stores the latest Redis stats for all clients
-func (ns *networkState) storeRedisStats(allStats map[redis.Key]*redis.RequestStats) {
-	storeUSMStats[redis.Key, *redis.RequestStats](
-		allStats,
-		ns.clients,
-		func(c *client) map[redis.Key]*redis.RequestStats { return c.redisStatsDelta },
-		func(c *client, m map[redis.Key]*redis.RequestStats) { c.redisStatsDelta = m },
-		func(prev, new *redis.RequestStats) { prev.CombineWith(new) },
-		ns.maxRedisStats,
-		stateTelemetry.redisStatsDropped.Inc,
-	)
-}
-
 func (ns *networkState) getClient(clientID string) *client {
 	if c, ok := ns.clients[clientID]; ok {
 		return c
